requirements.txt
attrs==19.3.0
bcrypt==3.1.7
certifi==2020.6.20
cffi==1.14.0
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
more-itertools==8.4.0
packaging==20.4
pluggy==0.13.1
py==1.9.0
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
pyparsing==2.4.7
pytest==5.4.3
python-dotenv==0.13.0
six==1.15.0
wcwidth==0.2.5
Werkzeug==1.0.1
